Extracting email footer text from templates
We currently have a set of nearly identical templates for each of our languages. In order to avoid having to make potential changes in multiple places in the future we will add a single template to the gimi-server project (where these templates are ultimately used) and break out the parts that vary between the different languages. One such part is the text at in the footer of many of our emails which we are extracting in this commit.