Remove unused backend languages (#607)
* Removing server translations for unsupported languages

When we retrieve translations in the backend we only do so for a configured group of languages (currently sv, nb, nl and en). When a translation for a language that doesn’t belong to this group is requested we fall back to the default language which currently is en. We can thus remove the other files to decrease confusion about why they exist and any other noise they may cause.

* Removing other translations for unsupported languages